--- examples/algorithms/always_defect.py
@@ -1,15 +1,0 @@
-from collections.abc import Sequence
-from typing import Self
-
-from project_dilemma.interfaces import Algorithm, Rounds
-
-
-class AlwaysDefect(Algorithm):
-    algorithm_id = 'always_defect'
-
-    def __init__(self, mutations: Sequence[Self]):
-        super().__init__(mutations)
-
-    @staticmethod
-    def decide(rounds: Rounds) -> bool:
-        return False
