+++ examples/config/mutable_complete_multiprocess_standard_simulation.toml
@@ -1,4 +1,4 @@
-simulation_id = "Complete MultiprocessStandardSimulation"
+simulation_id = "Mutable Complete MultiprocessStandardSimulation"
 algorithms_directory = "examples/algorithms/"
 nodes = [ { node_id = "always_cooperate", algorithm = { file = "simple.py", object = "AlwaysCooperate" } },
           { node_id = "always_defect", algorithm = { file = "simple.py", object = "AlwaysDefect" } },
@@ -8,6 +8,6 @@ nodes = [ { node_id = "always_cooperate", algorithm = { file = "simple.py", obje
           { node_id = "tit_for_two_tats", algorithm = { file = "tit_for_tat.py", object = "TitForTwoTats" } },
           { node_id = "two_tits_for_tat", algorithm = { file = "tit_for_tat.py", object = "TwoTitsForTat" } } ]
 simulation = { object = "MultiprocessStandardSimulation" }
-simulation_arguments = { rounds = 200 }
-rounds_output = "examples/rounds/complete_multiprocess_standard_simulation.json"
-simulation_output = "examples/results/complete_multiprocess_standard_simulation.json"
+simulation_arguments = { rounds = 200, round_mutations = true, mutations_per_mille = 100 }
+rounds_output = "examples/rounds/mutable_complete_multiprocess_standard_simulation.json"
+simulation_output = "examples/results/mutable_complete_multiprocess_standard_simulation.json"
