+++ examples/config/noisy_complete_multiprocess_standard_simulation.toml
@@ -1,12 +1,12 @@
 simulation_id = "Noisy Complete MultiprocessStandardSimulation"
 algorithms_directory = "examples/algorithms/"
-nodes = [ { node_id = "always_cooperate", algorithm = { file = "always_cooperate.py", object = "AlwaysCooperate" } },
-          { node_id = "always_defect", algorithm = { file = "always_defect.py", object = "AlwaysDefect" } },
-          { node_id = "grim_trigger", algorithm = { file = "grim_trigger.py", object = "GrimTrigger" } },
-          { node_id = "random_cooperation", algorithm = { file = "random_cooperation.py", object = "RandomCooperation" } },
+nodes = [ { node_id = "always_cooperate", algorithm = { file = "simple.py", object = "AlwaysCooperate" } },
+          { node_id = "always_defect", algorithm = { file = "simple.py", object = "AlwaysDefect" } },
+          { node_id = "grim_trigger", algorithm = { file = "grudge.py", object = "GrimTrigger" } },
+          { node_id = "random_cooperation", algorithm = { file = "simple.py", object = "RandomCooperation" } },
           { node_id = "tit_for_tat", algorithm = { file = "tit_for_tat.py", object = "TitForTat" } },
-          { node_id = "tit_for_two_tats", algorithm = { file = "tit_for_two_tats.py", object = "TitForTwoTats" } },
-          { node_id = "two_tits_for_tat", algorithm = { file = "two_tits_for_tat.py", object = "TwoTitsForTat" } } ]
+          { node_id = "tit_for_two_tats", algorithm = { file = "tit_for_tat.py", object = "TitForTwoTats" } },
+          { node_id = "two_tits_for_tat", algorithm = { file = "tit_for_tat.py", object = "TwoTitsForTat" } } ]
 simulation = { object = "MultiprocessStandardSimulation" }
 simulation_arguments = { rounds = 200, noise = true, noise_per_mille = 250}
 rounds_output = "examples/rounds/noisy_complete_multiprocess_standard_simulation.json"
