+++ examples/results/complete_multiprocess_standard_simulation.json
@@ -1,1 +1,1 @@
-{"always_cooperate": 2718, "always_defect": 4768, "random_cooperation": 3433, "grim_trigger": 3190, "tit_for_tat": 2670, "tit_for_two_tats": 2706, "two_tits_for_tat": 2679}
+{"always_cooperate": 2679, "grim_trigger": 3230, "always_defect": 4792, "random_cooperation": 3463, "tit_for_tat": 2682, "tit_for_two_tats": 2652, "two_tits_for_tat": 2691}
