+++ examples/results/complete_standard_simulation.json
@@ -1,1 +1,1 @@
-{"always_cooperate": 2727, "always_defect": 4824, "grim_trigger": 3190, "random_cooperation": 3351, "tit_for_tat": 2718, "tit_for_two_tats": 2733, "two_tits_for_tat": 2697}
+{"always_cooperate": 2727, "always_defect": 4764, "grim_trigger": 3220, "random_cooperation": 3359, "tit_for_tat": 2724, "tit_for_two_tats": 2745, "two_tits_for_tat": 2682}
