+++ examples/results/mutable_complete_multiprocess_standard_simulation.json
@@ -1,0 +1,1 @@
+{"always_cooperate": 3484, "always_defect": 3460, "grim_trigger": 3585, "random_cooperation": 3370, "tit_for_tat": 2511, "tit_for_two_tats": 2787, "two_tits_for_tat": 2763}
