+++ examples/results/mutable_complete_standard_simulation.json
@@ -1,0 +1,1 @@
+{"always_cooperate": 3212, "always_defect": 3465, "grim_trigger": 3501, "random_cooperation": 3712, "tit_for_tat": 2874, "tit_for_two_tats": 2685, "two_tits_for_tat": 2535}
