+++ examples/results/mutable_noisy_complete_multiprocess_standard_simulation.json
@@ -1,0 +1,1 @@
+{"always_cooperate": 2863, "always_defect": 2985, "random_cooperation": 2970, "grim_trigger": 3139, "tit_for_tat": 2592, "tit_for_two_tats": 2621, "two_tits_for_tat": 2621}
