+++ examples/results/mutable_noisy_complete_standard_simulation.json
@@ -1,0 +1,1 @@
+{"always_cooperate": 2957, "always_defect": 2911, "grim_trigger": 3147, "random_cooperation": 3107, "tit_for_tat": 2485, "tit_for_two_tats": 2683, "two_tits_for_tat": 2430}
