+++ examples/results/noisy_complete_multiprocess_standard_simulation.json
@@ -1,1 +1,1 @@
-{"always_cooperate": 2288, "two_tits_for_tat": 2731, "always_defect": 3492, "grim_trigger": 3211, "random_cooperation": 2937, "tit_for_tat": 2554, "tit_for_two_tats": 2422}
+{"always_cooperate": 2400, "always_defect": 3376, "grim_trigger": 3322, "random_cooperation": 3052, "tit_for_tat": 2551, "two_tits_for_tat": 2657, "tit_for_two_tats": 2424}
