+++ examples/results/noisy_complete_standard_simulation.json
@@ -1,1 +1,1 @@
-{"always_cooperate": 2394, "always_defect": 3543, "grim_trigger": 3112, "random_cooperation": 2904, "tit_for_tat": 2587, "tit_for_two_tats": 2482, "two_tits_for_tat": 2677}
+{"always_cooperate": 2405, "always_defect": 3360, "grim_trigger": 3117, "random_cooperation": 2887, "tit_for_tat": 2625, "tit_for_two_tats": 2436, "two_tits_for_tat": 2666}
