+++ src/project_dilemma/simulations/basic_simulation.py
@@ -17,8 +17,35 @@ from collections.abc import Sequence
 import random
 from typing import Optional
 
-from project_dilemma.interfaces import Node, Simulation, SimulationRounds
-from project_dilemma.simulations.utils import play_round
+from project_dilemma.interfaces import Node, Round, Rounds, Simulation, SimulationRounds
+
+
+def play_round(nodes: Sequence[Node],
+               rounds: Rounds,
+               *,
+               mutations_per_mille: int,
+               round_mutations: bool = False,) -> Round:
+    """run a round of prisoners dilemma with each node
+
+    :param nodes: nodes to run
+    :type nodes: Sequence[Node]
+    :param rounds: lsit of rounds
+    :type rounds: RoundList
+    :param mutations_per_mille: rate that mutations should occur per mille
+    :type mutations_per_mille: int
+    :param round_mutations: if mutations are enabled for rounds
+    :type round_mutations: bool
+    :return: round results
+    :rtype: Round
+    """
+    move: Round = {}
+    for node in nodes:
+        move[node.node_id] = node.algorithm.decide(rounds)
+
+        if round_mutations and random.randrange(0, 1000) < mutations_per_mille:
+            node.mutate()
+
+    return move
 
 
 class BasicSimulation(Simulation):
